+++ spec/components/brexit_cta_spec.rb
@@ -24,13 +24,13 @@ describe "BrexitCta", type: :view do
 
   it "shows standard brexit call to action" do
     render_component({ content_item: content_item })
-    assert_select ".gem-c-contextual-sidebar__brexit-heading", text: "Brexit"
-    assert_link_with_text_in(".gem-c-contextual-sidebar__brexit-text", "/brexit", "Check what you need to do")
+    assert_select ".gem-c-contextual-sidebar__heading", text: "Brexit"
+    assert_link_with_text_in(".gem-c-contextual-sidebar__text", "/brexit", "Check what you need to do")
   end
 
   it "shows standard brexit call to action for Welsh pages" do
     I18n.with_locale(:cy) { render_component({ content_item: welsh_content_item }) }
-    assert_select ".gem-c-contextual-sidebar__brexit-heading", text: "Brexit"
-    assert_link_with_text_in(".gem-c-contextual-sidebar__brexit-text", "/brexit.cy", "Gwiriwch beth mae angen i chi wneud")
+    assert_select ".gem-c-contextual-sidebar__heading", text: "Brexit"
+    assert_link_with_text_in(".gem-c-contextual-sidebar__text", "/brexit.cy", "Gwiriwch beth mae angen i chi wneud")
   end
 end