+++ spec/components/contextual_sidebar_spec.rb
@@ -36,8 +36,8 @@ describe "Contextual sidebar", type: :view do
             "links" => {
               "parent_taxons" => [
                 {
-                  "content_id" => "a4038b29-b332-4f13-98b1-1c9709e216bc",
-                  "title" => "Transport",
+                  "content_id" => "f441e630-2bba-4bbd-8c1b-22ae61cc00c2",
+                  "title" => "Importing",
                 },
               ],
             },
@@ -46,49 +46,10 @@ describe "Contextual sidebar", type: :view do
       },
     }
     render_component(content_item: content_item)
-    assert_select ".gem-c-contextual-sidebar__brexit-cta"
+    assert_select ".gem-c-contextual-sidebar__cta.gem-c-contextual-sidebar__cta--brexit"
   end
 
-  it "does not render Brexit CTA when we have a document id exception" do
-    content_item = {
-      "title" => "30 creative teams awarded up to £100,000 each for Festival UK* 2022 R&D project",
-      "content_id" => "c3752802-f091-43a9-ba90-33568fccf391",
-      "locale" => "en",
-      "links" => {
-        "taxons" => [
-          {
-            "content_id" => "e2ca2f1a-0ff3-43ce-b813-16645ff27904",
-            "title" => "Society and culture",
-            "phase" => "live",
-          },
-        ],
-      },
-    }
-    render_component(content_item: content_item)
-    assert_select ".gem-c-contextual-sidebar__brexit-cta", 0
-  end
-
-  it "does not render Brexit CTA when we have a document type exception" do
-    content_item = {
-      "title" => "Transport news story",
-      "content_id" => "3c402d90-fe77-49b9-a8aa-1800d4fc2b3d",
-      "locale" => "en",
-      "document_type" => "transaction",
-      "links" => {
-        "taxons" => [
-          {
-            "content_id" => "a4038b29-b332-4f13-98b1-1c9709e216bc",
-            "title" => "Transport",
-            "phase" => "live",
-          },
-        ],
-      },
-    }
-    render_component(content_item: content_item)
-    assert_select ".gem-c-contextual-sidebar__brexit-cta", 0
-  end
-
-  it "does not render Brexit CTA when we have a taxon exception" do
+  it "does not render Brexit CTA when not tagged to an allowed taxon" do
     content_item = {
       "title" => "Local transport news story",
       "content_id" => "5c82db20-7631-11e4-a3cb-005056011aef",
@@ -112,33 +73,6 @@ describe "Contextual sidebar", type: :view do
       },
     }
     render_component(content_item: content_item)
-    assert_select ".gem-c-contextual-sidebar__brexit-cta", 0
-  end
-
-  it "does not render Brexit CTA when locale is not 'en' or 'cy'" do
-    content_item = {
-      "title" => "السعي إلى إتمام القضاء على برنامج الأسلحة الكيميائية السوري",
-      "content_id" => "9ec092af-0f53-4a82-b4b0-9d016162ba01",
-      "locale" => "ar",
-      "links" => {
-        "taxons" => [
-          {
-            "content_id" => "3b4d6319-fcef-4637-b35a-e3df76321894",
-            "title" => "Local transport",
-            "phase" => "live",
-            "links" => {
-              "parent_taxons" => [
-                {
-                  "content_id" => "a4038b29-b332-4f13-98b1-1c9709e216bc",
-                  "title" => "Transport",
-                },
-              ],
-            },
-          },
-        ],
-      },
-    }
-    render_component(content_item: content_item)
-    assert_select ".gem-c-contextual-sidebar__brexit-cta", 0
+    assert_select ".gem-c-contextual-sidebar__cta.gem-c-contextual-sidebar__cta--brexit", 0
   end
 end