+++ spec/features/contextual_navigation_spec.rb
@@ -5,7 +5,6 @@ describe "Contextual navigation" do
     given_theres_a_page_with_brexit_taxon
     and_i_visit_that_page
     and_i_see_the_brexit_contextual_breadcrumbs
-    and_i_see_the_brexit_call_to_action
   end
 
   scenario "There is a page tagged to the brexit business child taxon" do
@@ -59,6 +58,7 @@ describe "Contextual navigation" do
     given_theres_a_travel_advice_page
     and_i_visit_that_page
     then_i_see_parent_breadcrumbs
+    and_i_see_the_covid_booster_call_to_action
   end
 
   scenario "The page has curated related items" do
@@ -94,7 +94,6 @@ describe "Contextual navigation" do
     then_i_see_the_step_by_step
     and_the_step_by_step_header
     and_i_do_not_see_the_brexit_contextual_breadcrumbs
-    and_i_see_the_brexit_call_to_action
   end
 
   scenario "It's a HTML Publication with a parent with breadcrumbs" do
@@ -446,7 +445,7 @@ describe "Contextual navigation" do
   end
 
   def and_i_see_the_other_step_by_step_as_an_also_part_of_list
-    within ".gem-c-step-nav-related:last-child" do
+    within ".gem-c-step-nav-related:last-of-type" do
       expect(page).to have_content("Also part of")
       expect(page).to have_content("PRIMARY STEP BY STEP - NOT INTERACTING WITH")
       expect(page).not_to have_content("PRIMARY STEP BY STEP - INTERACTING WITH")
@@ -454,7 +453,7 @@ describe "Contextual navigation" do
   end
 
   def and_i_dont_see_the_secondary_step_by_step_in_the_also_part_of_list
-    within ".gem-c-step-nav-related:last-child" do
+    within ".gem-c-step-nav-related:last-of-type" do
       expect(page).not_to have_content("SECONDARY STEP BY STEP")
     end
   end
@@ -530,16 +529,16 @@ describe "Contextual navigation" do
     and_i_see_the_brexit_contextual_breadcrumbs(brexit_individuals_taxon)
   end
 
-  def and_i_see_the_brexit_call_to_action
+  def and_i_see_the_covid_booster_call_to_action
     within ".gem-c-contextual-sidebar" do
-      expect(page).to have_selector(".gem-c-contextual-sidebar__brexit-cta")
-      expect(page).to have_css(".gem-c-contextual-sidebar__brexit-heading", text: I18n.t("components.related_navigation.transition.title"))
+      expect(page).to have_selector(".gem-c-contextual-sidebar__cta.gem-c-contextual-sidebar__cta--covid")
+      expect(page).to have_css(".gem-c-contextual-sidebar__heading", text: I18n.t("components.related_navigation.covid_booster.title"))
     end
   end
 
   def and_i_do_not_see_the_brexit_call_to_action
     within ".gem-c-contextual-sidebar" do
-      expect(page).not_to have_selector(".gem-c-contextual-sidebar__brexit-cta")
+      expect(page).not_to have_selector(".gem-c-contextual-sidebar__cta.gem-c-contextual-sidebar__cta--brexit")
     end
   end